feat: spray can limit usage
Probably not a good idea to depend on people to set the correct durability for a spray can when creating the item. For now, we can handle it this way until otherwise.

Although, there may not be much point in worrying about durability since there's a limit on the number of tags a person can place, which would just prevent them from using the item.